updated use of future packages
what seemed to happening is that future was creating background sessions or "workers" but doesn't seem to automatically close these sessions so RStudio doesn't know what to next because it has so many session open without clear direction of what to do with them.

Part of the weirdness of future creating "sessions" is that all loaded packages are not automatically loaded within each session... I tried using their recommended by using

.options = furrr_options(packages = loadedNamespaces())

But this cases untraceable errors and weird issues with select and map inside the analysis function... But using the load_packages() function and options(.) inside the future_map(.) resolved the issue and ran without issue. I'm not sure why they recommend against using library(.) inside future_map(.) and I since it worked I just went with it...

One funky part of using the future packages is that I had to explicitly write in a "reset" after each use to clear the background sessions and ensure that the next use of future works as intended. Again, not sure why this isn't being done automatically.