re-worked caching
Caching is now done explicitly to through a "cache.file" parameter to each .Rmd file. The explicit cache is only useful to make the functions run faster, and NOT to act as a true "cache." This can easily be modified, but the within function cache helps to ensure previous erroneous runs to these functions don't through errors in new calls.