revised document printing
So I think ?I actually got the bugs fixed for the revised way documents are generated and printed out. The one caveat is that the large summary tables are not printed out in the word document due to unresolvable errors with the officer package the it's dependence on xlm printing. but the PDF works.